labs/BEW-1.2-Forms-Homework/requirements.txt
click==7.1.2
Flask==1.1.2
Flask-SQLAlchemy==2.4.4
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
python-dotenv==0.15.0
six==1.15.0
SQLAlchemy==1.3.23
SQLAlchemy-Utils==0.36.8
Werkzeug==1.0.1
